% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri_functions.R
\name{.browse}
\alias{.browse}
\alias{carbon-browse}
\title{open $uri to in browser window}
\usage{
.browse(self, private)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}
}
\description{
opens the uri address in \link[=carbon-uri]{$uri} in the browser window.
}
\examples{
x <- carbon$new('x <- 1')
if(interactive())
 x$browse()

}
\seealso{
\link[=carbon]{carbon}
}
