% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_functions.R
\name{.chromeOptions}
\alias{.chromeOptions}
\alias{carbon-chrome}
\alias{.chrome_start}
\alias{.chrome_stop}
\title{chrome webdriver functions}
\usage{
.chromeOptions(self, private)

.chrome_start(self, private)

.chrome_stop(self, private)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}
}
\description{
Functions to open,close,set chrome webdriver
}
\seealso{
\link[carbonate:carbon]{carbon}, \link[wdman:chrome]{chrome}
}
