% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_functions.R
\name{.firefoxOptions}
\alias{.firefoxOptions}
\alias{carbon-gecko}
\title{gecko webdriver functions}
\usage{
.firefoxOptions(self, private)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}
}
\description{
Functions to open,close,set gecko webdriver
}
\seealso{
\link[=carbon]{carbon}, \link[wdman:gecko]{gecko}
}
