% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_functions.R
\name{.start}
\alias{.start}
\alias{.stop}
\alias{.stop_all}
\title{RSelenium webdriver functions}
\usage{
.start(self, private, eCap)

.stop(self, private)

.stop_all(self, private)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}

\item{eCap}{Rselenium options}
}
\description{
Functions to open,stop, stop_all RSelenium sessions
}
