% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonate.R
\name{.carbonate}
\alias{.carbonate}
\alias{carbon-carbonate}
\title{carbonate script}
\usage{
.carbonate(self, private, file, code, rD)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}

\item{file}{character, name of file to save image as}

\item{code}{character, lines of script to make carbon image from}

\item{rD}{RSelenium driver}
}
\value{
image object
}
\description{
Main function of the package that invokes RSelenium to open
a browser to the carbon.js uri, create an image and download the file.
}
\examples{
if(interactive()){
 x <- carbon$new('x <- 1')
 x$carbonate()
 }
}
\seealso{
\link[magick:editing]{editing}, \link[carbonate:carbon]{carbon}
}
