% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri_functions.R
\name{.encode}
\alias{.encode}
\alias{carbon-encode}
\title{encode uri for carbon.js}
\usage{
.encode(self, private, URL, reserved, repeated)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}

\item{URL}{character, uri string to url encode}

\item{reserved}{boolean, should ‘reserved’ characters be encoded?}

\item{repeated}{boolean, should apparently already-encoded URLs be encoded again?}
}
\value{
character
}
\description{
encode uri string for carbon.js.
Augmented version of than \link[utils:URLencode]{URLencode}
}
\seealso{
\link[utils:URLencode]{URLencode}, \link[carbonate:carbon]{carbon}
}
