% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get_functions.R
\name{.get_windows_control_themes}
\alias{.get_windows_control_themes}
\alias{carbon-get-fields}
\alias{.get_font_families}
\alias{.get_templates}
\title{carbon get functions}
\usage{
.get_windows_control_themes(self, private)

.get_font_families(self, private)

.get_templates(self, private)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}
}
\description{
function that gets the available list of aesthetic values of fields
}
\seealso{
\link[=carbon]{carbon}
}
