% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri_functions.R
\name{.options}
\alias{.options}
\alias{carbon-options}
\title{concatenate the carbon options to a string}
\usage{
.options(self, private, code)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}

\item{code}{character, script to embbed into the uri}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
combine all the carbon options into a carbon.js valid string
}
\seealso{
\link[=carbon]{carbon}
}
