% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{reprex_to_carbon_browser}
\alias{reprex_to_carbon_browser}
\alias{reprex_to_carbon_download}
\title{Creates a Reprex of selected code (or from input) renders it and opens or downloads the Carbon}
\usage{
reprex_to_carbon_browser(..., html_preview = FALSE, style = TRUE,
  url_only = FALSE)

reprex_to_carbon_download(..., file = "carbon_reprex.png",
  html_preview = FALSE, style = TRUE, url_only = FALSE)
}
\arguments{
\item{...}{arguments passed to \link[reprex:reprex]{reprex::reprex}}

\item{html_preview}{show also reprex html result? Default is false}

\item{style}{should the code be styled? Default is True}

\item{url_only}{if only the URL should be returned without opening the browser}
}
\value{
the URL to the Carbon (invisible) or the filename (invisible)
}
\description{
Creates a Reprex of selected code (or from input) renders it and opens or downloads the Carbon
}
\section{Functions}{
\itemize{
\item \code{reprex_to_carbon_download()}: downloads the picture

}}
\examples{
\dontrun{
reprex_to_carbon_browser(input = c("   x=1:  10", "   x"))
# copy the following code using STRG + C or similar:
  a <- 1:10
  b <- rnorm(10)
  cbind(a, b)
  plot(a, b)

# Then run this
reprex_to_carbon_browser()
}
}
