% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get_functions.R
\name{.set_template}
\alias{.set_template}
\alias{carbon-set-fields}
\alias{.set_window_control_theme}
\alias{.set_font_family}
\title{carbon set functions}
\usage{
.set_template(self, private, template)

.set_window_control_theme(self, private, theme)

.set_font_family(self, private, family)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}

\item{template}{character, name of template to set \link[=carbon-fields]{$template} to}

\item{theme}{character, theme to set \link[=carbon-fields]{$theme} to}

\item{family}{character, font family to set \link[=carbon-fields]{$font_family} to}
}
\description{
function that set the aesthetics of the carbon image output
}
\seealso{
\link[=carbon]{carbon}
}
