% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri_functions.R
\name{.tiny}
\alias{.tiny}
\alias{carbon-tinyurl}
\title{convert carbon uri to tinyurl}
\usage{
.tiny(self, private, clip = FALSE)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}

\item{clip}{boolean, write tinyurl to clipboard, Default: FALSE}
}
\value{
character
}
\description{
convert carbon uri to tinyurl with option to attach to clipboard.
}
\details{
If clip is set to TRUE \link[clipr:write_clip]{write_clip} will put the
tinyurl on the clipboard.
}
