% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri_functions.R
\name{.uri}
\alias{.uri}
\alias{carbon-uri}
\title{URI constructor}
\usage{
.uri(self, private, code)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}

\item{code}{character, script to embbed into the uri}
}
\description{
Function that is called in a carbon object to construct valid uri to send to browser
}
\examples{
if(interactive()){

x <- carbon$new('x <- 1')

# populate from self$code
x$uri()

# enter manually
x$uri(code = 'x <- y + 3')

}
}
\seealso{
\link[carbonate:carbon]{carbon}
}
