% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\docType{data}
\name{carbon}
\alias{carbon}
\title{carbon object}
\description{
The Carbon generator creates a new \code{Carbon}-object, which is the class containing
all the app logic. The class is based on the \link[R6:R6Class]{R6} OO-system and
is thus reference-based with methods and data attached to each object, in
contrast to the more well known S3 and S4 systems.
}
\section{Initialization}{

A new 'Carbon'-object is initialized using the \code{new()} method on the generator:

\tabular{l}{
\code{x <- carbon$new(code = clipr::read_clip())}
}
}

\section{Carbon Fields}{

Description of fields of the R6 object that can be set by the user can be found
in the following \link[carbonate:.carbon_fields]{page}.
}

\section{Main Methods}{

\tabular{ll}{
\link[carbonate:.carbonate]{$carbonate} \tab Using RSelenium fetch the carbon image output \cr
\link[carbonate:.browse]{$browse} \tab open \link[carbonate:.uri]{$uri} in the browser \cr
}
}

\section{Aesthetic Methods}{

\tabular{ll}{
\link[carbonate:.set_template]{$set_template} \tab set $template \cr
\link[carbonate:.get_templates]{$get_templates} \tab get a list of possible templates \cr
\link[carbonate:.set_window_control_theme]{$set_window_control_theme} \tab  set $windows_control_theme \cr
\link[carbonate:.get_windows_control_themes]{$get_windows_control_themes} \tab get a list of possible window control themes\cr
\link[carbonate:.set_font_family]{$set_font_family} \tab set $font_family\cr
\link[carbonate:.get_font_families]{$get_font_families} \tab get a list of possible fonts \cr
}
}

\section{URI Methods}{

\tabular{ll}{
\link[carbonate:.uri]{$uri} \tab construct valid carbon.js uri \cr
\link[carbonate:.options]{$options} \tab return all current carbon options\cr
\link[carbonate:.encode]{$encode} \tab URL encode a string for the $uri  \cr
}
}

\section{Webdriver Methods}{

\tabular{ll}{
\link[carbonate:.chromeOptions]{$chromeOptions} \tab construct a chromeOptions object \cr
\link[carbonate:.chrome_start]{$chrome_start} \tab start a chrome session \cr
\link[carbonate:.chrome_stop]{$chrome_stop} \tab stop a chrome session \cr
\link[carbonate:.start]{$start} \tab start a RSelenium session \cr
\link[carbonate:.stop]{$stop} \tab stop a RSelenium session \cr
\link[carbonate:.stop_all]{$stop_all} \tab stop all active RSelenium sessions \cr
}
}

\keyword{datasets}
