% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_functions.R
\name{.chromeOptions}
\alias{.chromeOptions}
\alias{carbon-chrome}
\title{chrome webdriver functions}
\usage{
.chromeOptions(self, private)
}
\arguments{
\item{self}{carbon self object}

\item{private}{carbon private object}
}
\description{
Functions to open,close,set chrome webdriver
}
\seealso{
\link[=carbon]{carbon}, \link[wdman:chrome]{chrome}
}
